+++ server/config.js
@@ -95,18 +95,7 @@ const config = {
   // },
 
   // Databases:
-  db: {
-    sqlite: {
-      database: './db/db.sqlite',
-    },
-    mysql: {
-      host: 'server domain name',
-      connectionLimit: 'number of simultaneous connections (number)',
-      database: 'database name',
-      user: 'username with the privileges to access to the database',
-      password: 'username password',
-    },
-  },
+  // See the file './.env.js'
 
   // Token
   // (lifetime is expressed in seconds)