+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   * Extended the support to MySQL/MariaDB databases,
   * Added the support of oauth tokens,
   * Added the option to revoke an access token,
+  * Added an hidden environment file to save the database credentials,
   * ...,